chore(ci): switch from flake8 to ruff
Switch from using [flake8](https://pypi.org/project/flake8) to [ruff](https://pypi.org/project/ruff) as the primary linter for the project. Ruff is orders of magnitude faster than alternative tools while integrating more functionality behind a single, common interface. This should help speed up linting for the project considerably and allow for much more sophisticated linting options and configurations. As a consequence of that, the following changes are included as part of this migration:

- Removal of flake8 from the project’s dependencies and pre-commit hooks.
- Addition of `ruff` to the project's dependencies and pre-commit hooks.
- Activation of additional linting rules supported by `ruff` and the accompanying changes on the source code to conform to these rules.
- Upgrading of existing dependencies.
- Removal of `tox.ini`. Tox configuration is now included in the `pyproject.toml` file. All the project configuration is now housed in one file. This is in line with [PEP 621](https://peps.python.org/pep-0621) guidelines.
- Minor fixes and improvements.