feat: add `UploadMetadataFactory` core interface
This new core interface defines a single operation `new_upload_meta` whose resposibility is to create and return new `UploadMetadata` instances. This respobility was previously defined as part of the `MetadataSink` interface but has now been moved to this new interface.

Included are refactors to the existing implementations to conform to the new changes.