feat: add db instances config support
Add a new settings initializer to load SQL database instances
configurations. The new new database instances configurations allow for
arbitrary databases to be loaded to the application through the configuration
as long right the corresponding database drivers are already installed.

Included is also a new property, `has_secrets` on the
`app.lib.config.SettingIntializer` interface that can be used to
indicate whether the value of a setting has secrets or is confidential.
This signals that the value should be handled with special care to
prevent accidental exposure.