feat: add env variables substitution from config
With this addition, config files (both toml and yaml) can contain environment
variables which are then substituted to their actual values before the file is
parsed and the `SettingInitializer`s are run.

Included is also the addition of the config format selector to allow the
correct config format to be specified when the automatic selector fails.