chore: refactor ETLProtocol to use single terminal values
Rename the `ETLProtocol` properties `metadata-sinks` and
`metadata_sources` to `metadata_consumer` and `metadata_supplier`
respectively. Also change the `ETLProtocol` interface properties
`metadata_consumer` and `metadata_supplier` to return single values
instead of an interable of values.

Included also are accompanying changes to implementing classes and
sub-interfaces to comply to the new interface.