feat: first draft of a concurrent runtime
Add an initial implementation of concurrency on the runtime. This
implementation uses threads to achieve concurrency but only for IO bound
tasks. In the future, more options might be explored, such as using
asycio or multiprocessing libraries.

Included are also fixes and minor improvements to existing code to make
it more suited to work in a concurrent environment.