feat: add source meta properties to DataSource and DataSink
Add the properties `data_sink_meta` and `data_source_meta` to the
`DataSink` and `DataSource` interfaces respectively. This changes are
also cascaded down to the implementing classes.

Included also is a rewrite of some of parts of the IDR Client components
to utilize the new properties and other minor improvements and fixes.