feat: improve runtime setup function
Improve the `runtime.setup.setup` funtion to make it adequate as the runtime initializer.
This involves refactor the function to check if the application
dispacther has been set and if not, setting one before proceding with
the rest of the initializing.
Also move the `PreConfigSignal` and `PostConfigSignal` sending from the
main funtion to this function. This will ensure that this signals are
always fired even if the runtime is used as a library.